fix(chore): rename the file to match the actual function #25
